move book-settings
