+++ apps/client/src/pages/book-settings/api/change-book-title.graphql
